WESTON-CUBE: TOUCH-PLAYER: update
WESTON-CUBE: BO not backup-ed
Formerly BO was not backup-ed in the structure and then not destroyed
accordingly.
Still there is a file descriptor leakage similar to kmscube.

TOUCH-PLAYER: put window on full screen
In case we ask to don't have the playback on fullscreen,
the window need to be on fullscreen to avoid lost of focus.

Signed-off-by: Pierre-Yves MORDRET <pierre-yves.mordret@st.com>
Signed-off-by: Christophe Priouzeau <christophe.priouzeau@st.com>