dcmipp-isp-ctrl: initial commit of the DCMIPP ISP Control
dcmipp-isp-ctrl is an application which demonstrates some of
the features of the DCMIPP ISP blocks such as:
 - ISP statistics retrieval
 - ISP parameters injection

Expected usage of the application is using -i and -g parameters
in order to perform settings of the illuminance and auto-exposure
to reach a targeted level by both adjusting the DCMIPP exposure
block and the sensor gain.

It demonstrate as well the control of the contrast enhancement block
via the -c option.

Signed-off-by: Alain Volmat <alain.volmat@foss.st.com>